+++ src/lib/shared/types.ts
@@ -141,9 +141,18 @@ export type Pet = {
 	name: string;
 	levels: PetLevel[];
 };
+export type Guide = {
+	/** ID of the guide in the `army_guides` table */
+	id: number;
+	armyId: number;
+	textContent: string | null;
+	youtubeUrl: string | null;
+	createdTime: string;
+	updatedTime: string;
+};
 
 /**
- * The following `NewFoo` types are the bare minimum data required for army creation/saving
+ * The following `SaveFoo` types are the bare minimum data required for army creation/saving
  */
 export type SaveUnit = {
 	id?: number;
@@ -163,6 +172,7 @@ export type SavePet = {
 	petId: number;
 	hero: HeroType;
 };
+export type SaveGuide = Optional<Omit<Guide, 'createdTime' | 'updatedTime' | 'armyId'>, 'id'>;
 
 /**
  * The following `ArmyFoo` types is the complete data `getArmy` will return for convenience
@@ -179,6 +189,10 @@ export type ArmyPet = (SavePet & Omit<Pet, 'levels'>) & {
 	/** ID of the pet in the `army_pets` table */
 	id: number;
 };
+export type ArmyGuide = SaveGuide & {
+	/** ID of the guide in the `army_guides` table */
+	id: number;
+};
 
 /** The bare minimum data required for army creation/saving */
 export type SaveArmy = {
@@ -190,14 +204,16 @@ export type SaveArmy = {
 	units: SaveUnit[];
 	equipment: SaveEquipment[];
 	pets: SavePet[];
+	guide: SaveGuide | null;
 };
 /** A complete saved army, as returned by `getArmy` */
-export type Army = Omit<SaveArmy, 'units' | 'equipment' | 'pets'> & {
+export type Army = Omit<SaveArmy, 'units' | 'equipment' | 'pets' | 'guide'> & {
 	/** ID is now required */
 	id: number;
 	units: ArmyUnit[];
 	equipment: ArmyEquipment[];
 	pets: ArmyPet[];
+	guide: ArmyGuide | null;
 	createdBy: number;
 	createdTime: string;
 	updatedTime: string;